pyproject.toml
[project]
name = "rapport-amiante"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai (>=1.66.0,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "pdfplumber (>=0.11.9,<0.12.0)"
]

[tool.poetry]
packages = [{include = "rapport_amiante", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.15.5,<0.16.0)"
]
